% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAscore.R
\name{DAscore}
\alias{DAscore}
\title{Calculate the differential abundance (DA) score and show the DA figure}
\usage{
DAscore(
  increase_metabolites,
  decrease_metabolites,
  all_metabolites,
  sort_plot = NA,
  min_measured_num = 2
)
}
\arguments{
\item{increase_metabolites}{the increased metabolites  that needed kegg id}

\item{decrease_metabolites}{the decreased metabolites that needed kegg id}

\item{all_metabolites}{the all metabolites that kegg id}

\item{sort_plot}{the method of pathway rank in the plot,Default rank by DA score.
if value is classification, then rank by pathway classification}

\item{min_measured_num}{the minimum measured metabolites that be ploted in a pathway,
Default the value is 2.}
}
\value{
Calculate the differential abundance (DA) score and show the DA figure
}
\description{
Calculate the differential abundance (DA) score and show the DA figure
}
\examples{
name <- c("C00022","C00024","C00031")
test <- DAscore(name[1],name[2],name,min_measured_num = 0,sort_plot = "classification")
test$result
test$p


}
