% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM.R
\name{DM}
\alias{DM}
\title{fold change, p value and vip of the data}
\usage{
DM(mydata, group)
}
\arguments{
\item{mydata}{the rows are metabolites, and the columns are the samples}

\item{group}{the group information of the sample}
}
\value{
a data frame of the result contains fold change,p value and vip
}
\description{
fold change, p value and vip of the data
}
\examples{
library(dplyr)
result <- DM(mydata,group)
}
