% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESEA.R
\name{ESEA}
\alias{ESEA}
\title{the ESEA analysis}
\usage{
ESEA(
  Ranks_all,
  minSize = 5,
  nPermSimple = 20000,
  gseaParam = 0.5,
  out = "Extended"
)
}
\arguments{
\item{Ranks_all}{the Named vector of compound-level stats or gene-level stats. Names should be KEGG ID and (or) gene symbol.}

\item{minSize}{Minimal size of a compound and(or) gene set to test. All pathways below the threshold are excluded.}

\item{gseaParam}{GSEA parameter value, all compound-level and(or) gene-level stats are raised to the power of 'gseaParam' before calculation of GSEA enrichment scores.}

\item{out}{The pathway type for gene or metabolite,or extended pathway included genes and metabolites,default is "Extended",alternative is "metabolite" and "gene"}
}
\value{
the ESEA result
}
\description{
the Extended-data Set Enrichment Analysis which includes gene and metabolite
}
\examples{
library(ggplot2)
library(dplyr)
result <- ESEA(sim.cpd.data,out="Extended")
}
