% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keggid2pathway.R
\name{keggid2pathway}
\alias{keggid2pathway}
\title{search the kegg pathway corresponding to kegg id}
\usage{
keggid2pathway(keggid)
}
\arguments{
\item{keggid}{the kegg id that be searched pathway}
}
\value{
test
}
\description{
search the kegg pathway corresponding to kegg id
}
\examples{
keggid <- c("C05984","C02494")
kegg_result <- keggid2pathway(keggid)
}
