% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaGeneCor.R
\name{MetaGeneCor}
\alias{MetaGeneCor}
\title{The correlation between metabolites and gene expression in a pathway}
\usage{
MetaGeneCor(metadata, genedata, pathwayid)
}
\arguments{
\item{metadata}{the metabolites data, the row is the metabolites, and the column is the sample}

\item{genedata}{the gene expression data, the row is the gene, and the column is the sample}

\item{pathwayid}{the interested pathway id}
}
\value{
test
}
\description{
The correlation between metabolites and gene expression in a pathway
}
\examples{
library(dplyr)
metadata <- readRDS("/Users/guituantuan/Desktop/projects/RNA-seq/ALL/result_v0612/11_key_metabolites/metabolites_data_A.rds") \%>\%
as.data.frame()
metadata <- metadata[1:10,]
metadata[1,1] <- "Glutamic acid"

result_keggid <- name2keggid(metadata$label) \%>\%
  tidyr::separate_rows(kegg_id,sep=";") \%>\%
  dplyr::filter(!is.na(kegg_id)) \%>\%
  dplyr::left_join(metadata,by=c("name"="label"))

metadata <- result_keggid \%>\%
  dplyr::select(-"name") \%>\%
  tibble::column_to_rownames("kegg_id")

genedata <- readRDS("/Users/guituantuan/Desktop/projects/RNA-seq/ALL/result_v0612/11_key_metabolites/20220612_RNA_ALL_VST_coding_filter.rds") \%>\%
  as.data.frame() \%>\%
  tibble::column_to_rownames("gene_id")

result <- MetaGeneCor(metadata,genedata,"hsa00630")
}
