% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetCox.R
\name{MetCox}
\alias{MetCox}
\title{The cox analysis about the metabolites}
\usage{
MetCox(dat)
}
\arguments{
\item{object}{A dataframe-like data object containing log-metabolite intensity values, with rows corresponding to samples, and the columns corresponding to the metablites and must containing time and status}
}
\value{
result
}
\description{
The cox analysis about the metabolites
}
\examples{
result <- MetCox(dat_surv)
result

}
