% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_alpha.R
\name{ML_alpha}
\alias{ML_alpha}
\title{feature selection using lasso or elastic network}
\usage{
ML_alpha(object, method = "lasso")
}
\arguments{
\item{object}{A dataframe-like data object containing log-metabolite intensity values, with columns corresponding to metabolites and must containing the group column, and the rows corresponding to the samples}

\item{method}{the feature selection method,default is "lasso", "elastic" is optional}
}
\value{
test
}
\description{
feature selection using lasso or elastic network
}
\examples{
library(dplyr)
meta_dat1 <- t(meta_dat) \%>\%
  as.data.frame() \%>\%
  dplyr::mutate(group=group)
result_ML_lasso <- ML_alpha(meta_dat1,method="lasso")
}
