% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_Boruta.R
\name{ML_Boruta}
\alias{ML_Boruta}
\title{feature selection in Boruta}
\usage{
ML_Boruta(mydata)
}
\arguments{
\item{mydata}{the data}
}
\value{
test
}
\description{
feature selection in Boruta
}
\examples{
library(dplyr)
mydata_t <- mydata \%>\%
  t() \%>\%
  as.data.frame()
# the group information must be tumor and normal
mydata_t$group <- group
result <- ML_Boruta(mydata_t)
}
