% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_logistic.R
\name{ML_logistic}
\alias{ML_logistic}
\title{The logistic regression to predict the data}
\usage{
ML_logistic(mydata, out_dir = "logistic/", seed = NULL, auc_cutoff = 0.8)
}
\arguments{
\item{mydata}{the data that input}

\item{out_dir}{the out directory}

\item{seed}{the seed}

\item{auc_cutoff}{the cut off of auc that be filtered}
}
\value{
test
}
\description{
The logistic regression to predict the data
}
\examples{
library(dplyr)
}
