% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_RF.R
\name{ML_RF}
\alias{ML_RF}
\title{feature selection in random forest}
\usage{
ML_RF(mydata, ylim_min = 0, seed = NULL)
}
\arguments{
\item{mydata}{the data}

\item{ylim_min}{the min ylim,default is 0}

\item{seed}{the seed}
}
\value{
test
}
\description{
feature selection in random forest
}
\examples{
library(dplyr)
mydata_t <- mydata \%>\%
  t() \%>\%
  as.data.frame()
# the group information must be tumor and normal
mydata_t$group <- group
result <- ML_RF(mydata_t)
}
