% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_xgboost.R
\name{ML_xgboost}
\alias{ML_xgboost}
\title{Feature selection using XGBoost}
\usage{
ML_xgboost(object)
}
\arguments{
\item{object}{A dataframe-like data object containing log-metabolite intensity values, with columns corresponding to metabolites and must containing the group column, and the rows corresponding to the samples}
}
\value{
test
}
\description{
Feature selection using XGBoost
}
\examples{
library(dplyr)

meta_dat1 <- t(meta_dat) \%>\%
  as.data.frame() \%>\%
  dplyr::mutate(group=group)
result_ML_xgboost <- ML_xgboost(meta_dat1)

result_ML_xgboost$p
result_ML_xgboost$feature_result

}
