% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlimma.R
\name{mlimma}
\alias{mlimma}
\title{the limma analysis for metabolism or transcriptome}
\usage{
mlimma(dat, group)
}
\arguments{
\item{dat}{the data to be analyzed, the column is the sample name, and row is the gene name or compound name}

\item{group}{the data's group}
}
\value{
test
}
\description{
the limma analysis for metabolism or transcriptome
}
\examples{
result <- mlimma(mydata,group)
}
