\name{mnet-internal}

\title{Internal mnet functions}

\alias{Process_mlVAR}
\alias{print.mlVAR_GC}
\alias{ExampleData}
\alias{dataFried}

\description{Internal mgm functions.}

\author{Jonas Haslbeck <jonashaslbeck@protonmail.com>}

\details{
  These are internal functions or datasets. Process_mlVAR() is a function that takes the output of two mlVAR-objects (from the mlVAR package) and extracts all relevant parameters and their group differences.
}

\value{List different internal functions, different return values.}

\keyword{internal}

