% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSEA.R
\name{MSEA}
\alias{MSEA}
\title{the MSEA (Metabolite Set Enrichment Analysis) result}
\usage{
MSEA(Ranks_all, minSize = 5, nPermSimple = 20000, gseaParam = 0.5)
}
\arguments{
\item{Ranks_all}{the Named vector of compound-level stats. Names should be the same as in 'pathways'}

\item{minSize}{Minimal size of a compound set to test. All pathways below the threshold are excluded.}

\item{gseaParam}{GSEA parameter value, all compound-level statis are raised to the power of 'gseaParam' before calculation of GSEA enrichment scores.}
}
\value{
the MSEA result
}
\description{
the MSEA (Metabolite Set Enrichment Analysis) result
}
\examples{
library(ggplot2)
library(dplyr)
result <- MSEA(sim.cpd.data)
}
