% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myscale.R
\name{myscale}
\alias{myscale}
\title{Scale the data}
\usage{
myscale(object, method = "log_zscore")
}
\arguments{
\item{object}{A dataframe-like data object containing raw metabolite intensity values, with rows corresponding to metabolites, and the columns corresponding to the samples}

\item{method}{default is "log_zscore","raw_zscore" is alternative}
}
\value{
test
}
\description{
Scale the data
}
\examples{
object_scale <- myscale(meta_dat, method="raw_zscore")
object_scale

}
