% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name2keggid.R
\name{name2keggid}
\alias{name2keggid}
\title{Change the compound name to kegg id}
\usage{
name2keggid(compound_name)
}
\arguments{
\item{compound_name}{the compound name needed to change}
}
\value{
test
}
\description{
Change the compound name to kegg id
}
\examples{
compound_name <- c("2-Hydroxybutyric acid","1-Methyladenosine","tt","2-Aminooctanoic acid")
name2keggid(compound_name)

}
