% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name2pathway.R
\name{name2pathway}
\alias{name2pathway}
\title{seek the metabolits' pathway}
\usage{
name2pathway(name)
}
\arguments{
\item{name}{the metabolites' name}
}
\value{
test
}
\description{
seek the metabolits' pathway
}
\examples{
name <- c("2-Hydroxybutyric acid","1-Methyladenosine","tt")
result_all <- name2pathway(name)
result <- result_all$name2pathway
pathway <- result_all$pathway
kegg_name <- result_all$kegg_name
}
