% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathGeneMeta.R
\name{PathGeneMeta}
\alias{PathGeneMeta}
\title{the pathway analysis of combination of gene and metabolites}
\usage{
PathGeneMeta(name, out = "name")
}
\arguments{
\item{name}{the gene and the metabolites' name}

\item{out}{the figure's name type, default is "name"}
}
\value{
test
}
\description{
the pathway analysis of combination of gene and metabolites
}
\examples{
name <- c("C15973","C16254","MDH1")
result <- PathGeneMeta(name)
}
