% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway2pathwayid.R
\name{pathway2pathwayid}
\alias{pathway2pathwayid}
\title{Change the pathway name to pathway id}
\usage{
pathway2pathwayid(pathwayname)
}
\arguments{
\item{pathwayname}{the pathway name,such as "Glycolysis / Gluconeogenesis"}
}
\value{
test
}
\description{
Change the pathway name to pathway id
}
\examples{
pathwayid <- pathway2pathwayid("Glycolysis / Gluconeogenesis")
}
