% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayinfo.R
\name{pathwayinfo}
\alias{pathwayinfo}
\title{the gene and the metabolite in the pathway}
\usage{
pathwayinfo(pathwayid)
}
\arguments{
\item{pathwayid}{the pathway id, hsa or name}
}
\value{
the genes and metabolites KEGG ID corresponding to pathway
}
\description{
the gene and the metabolite in the pathway
}
\examples{
result <- pathwayinfo("hsa00630")
result <- pathwayinfo("Glyoxylate and dicarboxylate metabolism")
}
