% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pCliTS.R
\name{pCliTS}
\alias{pCliTS}
\title{The clinical's time series analysis}
\usage{
pCliTS(clinical, marker)
}
\arguments{
\item{clinical}{the clinical marker of different time, and need included "time","group",maker; "high"and "low" are alternative.}
}
\value{
Figure
}
\description{
The clinical's time series analysis
}
\examples{
time_series_ALT <- pCliTS(clinical_index)
time_series_ALT

}
