% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pCorCliMet.R
\name{pCorCliMet}
\alias{pCorCliMet}
\title{Plot the correlation of clinical or gene expression data with metabolites data}
\usage{
pCorCliMet(
  other_data,
  metabolite_data,
  other_marker,
  metabolite_marker,
  method = "kendall"
)
}
\arguments{
\item{other_data}{all the clincial data, or the gene expression data}

\item{metabolite_data}{all the metabolite data,the metabolite's sample order needed to be same with clinical data or the gene expression data}

\item{other_marker}{the filtered clinical marker, or the gene expression marker}

\item{metabolite_marker}{the filtered metabolite marker}

\item{method}{the correlation method}
}
\value{
test
}
\description{
Plot the correlation of clinical or gene expression data with metabolites data
}
\examples{
result <- pCorCliMet(clinical,mydata,"CRP","1-Methyladenosine")
}
