% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pCorCliMetMantel.R
\name{pCorCliMetMantel}
\alias{pCorCliMetMantel}
\title{The correlation plot of clinical and metabolite using mantel test}
\usage{
pCorCliMetMantel(clinical_data, metabolite_data)
}
\arguments{
\item{clinical_data}{the clinical data that row is the sample, and the column is the clinical biomarker}

\item{metabolite_data}{the metabolite data that row is the sample, and the column is the metabolites biomarker}
}
\value{
test
}
\description{
The correlation plot of clinical and metabolite using mantel test
}
\examples{
mydata1 <- t(mydata)
clinical_data <- as.data.frame(t(clinical))
metabolite_data <- as.data.frame(mydata1)
p <- pCorCliMetMantel(clinical_data,metabolite_data)
}
