% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdnet.R
\name{pdnet}
\alias{pdnet}
\title{The important network of the metabolite and the gene expression}
\usage{
pdnet(diff_metabolite, diff_gene, nsize = 10)
}
\arguments{
\item{diff_metabolite}{the differential metabolite information, the name 's kegg id}

\item{diff_gene}{the differential genes information}

\item{nsize}{the desired number of nodes constrained to the resulting subgraph}
}
\value{
a figure
}
\description{
The important network of the metabolite and the gene expression
}
\examples{
load("/Users/guituantuan/Desktop/R_packages/MNet/R/test.data")
pdf("test.pdf")
pdnet(diff_metabolite,diff_gene,nsize=10)
dev.off()
}
