% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pESEA.R
\name{pESEA}
\alias{pESEA}
\title{The visualization of ESEA in a pathway}
\usage{
pESEA(
  pathway_name,
  Ranks_all,
  out = "Extended",
  gseaParam = 0.5,
  minSize = 5,
  ticksSize = 0.2
)
}
\arguments{
\item{pathway_name}{the pathway name for plot}

\item{Ranks_all}{the Named vector of compound-level and (or) gene-level stats. Names should be KEGG ID and(or) gene-level}

\item{out}{The pathway type for gene or metabolite,or extended pathway included genes and metabolites,default is "Extended",alternative is "metabolite" and "gene"}

\item{gseaParam}{GSEA parameter value, all compound-level and(or) gene-level stats are raised to the power of 'gseaParam' before calculation of GSEA enrichment scores.}

\item{minSize}{Minimal size of a compound and(or) gene set to test. All pathways below the threshold are excluded.}

\item{ticksSize}{width of vertical line corresponding to a compound or gene(default: 0.2)}
}
\value{
visualization for the pathway's ESEA
}
\description{
The visualization of ESEA in a pathway
}
\examples{
result <- pESEA("Butanoate metabolism",sim.cpd.data,out="metabolite")
result

}
