% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pHeatmap.R
\name{pHeatmap}
\alias{pHeatmap}
\title{the heatmap of differential metabolites in the data}
\usage{
pHeatmap(
  mydata_diff,
  group,
  fontsize_row = 5,
  fontsize_col = 4,
  clustering_method = "complete",
  clustering_distance_cols = "correlation"
)
}
\arguments{
\item{mydata_diff}{the data of differential metabolites' value,the row is
metabolites, and the column is the samples' names.}

\item{group}{character vector,the data's group information, must tumor and normal}

\item{fontsize_row}{numeric, the value of row's fontsize, default is 5}

\item{fontsize_col}{numeric, the value of column's fontsize, default is 4}

\item{clustering_method}{the method that cluster,default is complete}

\item{clustering_distance_cols}{the distance for cluster,default is correlation}
}
\value{
test
}
\description{
the heatmap of differential metabolites in the data
}
\examples{
p_heatmap <- pHeatmap(mydata,group)
}
