% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pNetCor.R
\name{pNetCor}
\alias{pNetCor}
\title{The correlation network's extraction}
\usage{
pNetCor(
  dat_meta,
  dat_gene,
  cor_method = "pearson",
  cor_threshold = 0.9,
  p_threshold = 0.01
)
}
\arguments{
\item{dat_meta}{the metabolite data, the row is the metabolite's KEGGID, and the column is the sample}

\item{dat_gene}{the gene expression data, the row is the gene symbol, and the column is the sample}

\item{cor_method}{specifies the type of correlations to compute. Spearman correlations are the Pearson linear correlations computed on the ranks of non-missing elements, using midranks for ties.}

\item{cor_threshold}{the cutoff of correlation coefficient}

\item{p_threshold}{the cutoff of correlation p value}
}
\value{
a figure and a data frame
}
\description{
The correlation network's extraction
}
\examples{
a <- pNetCor(meta_dat,gene_dat)
}
