% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pPathview.R
\name{pPathview}
\alias{pPathview}
\title{the pathview of differential metabolites}
\usage{
pPathview(cpd.data, gene.data = NULL)
}
\arguments{
\item{cpd.data}{character vector,the metabolites' value, log2(fold change)}

\item{gene.data}{default is NULL, the gene expression data, log2(fold change)}
}
\value{
test
}
\description{
the pathview of differential metabolites
}
\examples{
dir.create("test")
setwd("test")
kegg_id <- c("C02494","C03665","C01546","C05984","C14088","C00587")
value <- c(-0.3824620,0.1823628,-1.1681486,0.5164899,1.6449798,-0.7340652)
names(value) <- kegg_id
cpd.data <- value
pPathview(cpd.data)
}
