% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pVolcano.R
\name{pVolcano}
\alias{pVolcano}
\title{Plot the volcano figure}
\usage{
pVolcano(mydata, foldchange_threshold = 1.5, p_threshold = 0.05)
}
\arguments{
\item{mydata}{the result of function "differential_metabolites",have column "Name","Fold_change", "Padj_t"}

\item{foldchange_threshold}{the fold change, default is 1.5,1/1.5}

\item{p_threshold}{the p value cutoff, default is 0.05}
}
\value{
test
}
\description{
Plot the volcano figure
}
\examples{
diff_result <- DM(2**meta_dat,group)
p_volcano <- pVolcano(diff_result,foldchange_threshold=1.5)
p_volcano

}
