% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_metabolite_gene.R
\name{reaction_metabolite_gene}
\alias{reaction_metabolite_gene}
\title{get the reaction from the gene and the metabolite}
\usage{
reaction_metabolite_gene(genesymbol, metabolite)
}
\arguments{
\item{genesymbol}{the gene symbol}

\item{metabolite}{the metabolite's kegg id}
}
\value{
test
}
\description{
get the reaction from the gene and the metabolite
}
\examples{
genesymbol <- "SLCO1B1"
metabolite <- "C00288"
result <- reaction_metabolite_gene(genesymbol,metabolite)
}
