% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survCli.R
\name{survCli}
\alias{survCli}
\title{The OS,RFS and EFS's survival analysis}
\usage{
survCli(clinical_survival)
}
\arguments{
\item{clinical_survival}{the data.frame must have the column name of OS_status, OS_time, group;
the column name of RFS_status,RFS_time,EFS_status,EFS_time is alternative}
}
\value{
test
}
\description{
The OS,RFS and EFS's survival analysis
}
\examples{
p <- survCli(clinical_survival)
p$p_OS
p$p_RFS
p$p_EFS
}
