% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survMet.R
\name{survMet}
\alias{survMet}
\title{The metabolites' survival analysis and visualization}
\usage{
survMet(
  dat,
  metabolites,
  cluster_method = "mean",
  out_dir = "survival/metabolites/"
)
}
\arguments{
\item{dat}{A dataframe-like data object containing log-metabolite intensity values, with rows corresponding to samples, and the columns corresponding to the metablites and must containing time and status}

\item{metabolites}{the selected metabolites that used to classify the samples}

\item{cluster_method}{"mean" or "median",defalult is "mean"}

\item{out_dir}{the output directory,default is "survival/metabolites"}
}
\value{
Figures
}
\description{
automatic classify the samples by each selected metabolites by mean or median, and then plot the survival
}
\examples{
metabolites <- c("C03819","C02918","C03916")
survMet(dat_surv,metabolites,cluster_method="mean",out_dir="result/survival/")
}
