% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp.R
\name{tomicsvis}
\alias{tomicsvis}
\title{MNet shiny app start function.}
\usage{
tomicsvis()
}
\value{
Shinyapp: MNet shiny app.
}
\description{
MNet shiny app start function.
}
\examples{
library(MNet)
# MNet::shinyapp()

}
