% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSMfuzz.R
\name{TSMfuzz}
\alias{TSMfuzz}
\title{The time series analysis}
\usage{
TSMfuzz(mydata, out_dir = "time_series_mfuzz/", range = c(4, 12))
}
\arguments{
\item{mydata}{the data that needed to be ranked by the time}

\item{out_dir}{the result's directory}

\item{range}{the metabolites's class's range,default is from 4 to 12}
}
\value{
test
}
\description{
The time series analysis
}
\examples{
TSMfuzz(mydata,out_dir="./",range=c(4,12))
}
