% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSSupraHex.R
\name{TSSupraHex}
\alias{TSSupraHex}
\title{The time series analysis using supraHex}
\usage{
TSSupraHex(
  mydata,
  newdata = NULL,
  out_dir = "time_series_supraHex/",
  rect.grid = c(1, 5)
)
}
\arguments{
\item{mydata}{the data that needed to be ranked by the time}

\item{newdata}{the data that use the model}

\item{out_dir}{the result's directory}

\item{rect.grid}{the}
}
\value{
test
}
\description{
The time series analysis using supraHex
}
\examples{
library(dplyr)
TSSupraHex(mydata,out_dir="time_series_supraHex/")
}
