% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgr.R
\name{xgr}
\alias{xgr}
\title{Plot the figure using XGR}
\usage{
xgr(metabolites_keggid, database, p_cutoff = 0.05, noverlap_cutoff = 0)
}
\arguments{
\item{metabolites_keggid}{the metabolites's keggid}

\item{database}{the database used}

\item{p_cutoff}{p_cutoff used to declare the significant terms. By default, it is set to 0.05}

\item{noverlap_cutoff}{noverlap_cutoff used to declare the number of overlap. By default, it is set to 0}
}
\value{
test
}
\description{
Plot the figure using XGR
}
\examples{
library(dplyr)
kegg_pathway_filter <- kegg_pathway \%>\%
dplyr::filter(!is.na(pathway_type)) \%>\%
  dplyr::select(ENTRY,PATHWAY)

kegg_id_need <- c("C05984","C02494")
xgr_result <- xgr(kegg_id_need,kegg_pathway_filter,p_cutoff=1.1,noverlap_cutoff=0)
xgr_result$output
xgr_result$gp
}
