% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keggid2pathway.R
\name{keggid2pathway}
\alias{keggid2pathway}
\title{The kegg pathway corresponding to kegg id}
\usage{
keggid2pathway(keggid)
}
\arguments{
\item{keggid}{the KEGG ID}
}
\value{
result
}
\description{
The kegg pathway corresponding to kegg id
}
\examples{
keggid <- c("C05984","C02494")
kegg_result <- keggid2pathway(keggid)
kegg_result

}
