% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_impute.R
\name{meta_impute}
\alias{meta_impute}
\title{Impute the NA in data}
\usage{
meta_impute(object, method, k = k)
}
\arguments{
\item{object}{object A dataframe-like data object containing raw metabolite intensity values, with rows corresponding to metabolites, and the columns corresponding to the samples}

\item{method}{the method used for imputed data, default is min_0.1. You can choose "mean","median","min","knn"}

\item{k}{the k value for knn imputation}
}
\value{
a data frame
}
\description{
Impute the NA in data
}
\examples{
library(MNet)

}
