% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_alpha.R
\name{ML_alpha}
\alias{ML_alpha}
\title{feature selection in lasso or elastic network}
\usage{
ML_alpha(mydata, method = "lasso")
}
\arguments{
\item{mydata}{the data}

\item{method}{the feature selection method,default is "lasso", "elastic" is optional}
}
\value{
test
}
\description{
feature selection in lasso or elastic network
}
\examples{
library(dplyr)
mydata_t <- mydata \%>\%
  t() \%>\%
  as.data.frame()
# the group information must be tumor and normal
mydata_t$group <- group
result <- ML_alpha(mydata_t)
}
