% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_Boruta.R
\name{ML_Boruta}
\alias{ML_Boruta}
\title{feature selection using Boruta}
\usage{
ML_Boruta(object)
}
\arguments{
\item{object}{A dataframe-like data object containing log-metabolite intensity values, with columns corresponding to metabolites and must containing the group column, and the rows corresponding to the samples}
}
\value{
test
}
\description{
feature selection using Boruta
}
\examples{
library(dplyr)
meta_dat1 <- t(meta_dat) \%>\%
  as.data.frame() \%>\%
  dplyr::mutate(group=group)
result_ML_Boruta <- ML_Boruta(meta_dat1)
}
