% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_logistic.R
\name{ML_logistic}
\alias{ML_logistic}
\title{Feature selection using logistic regression}
\usage{
ML_logistic(object, out_dir = "logistic/", seed = NULL, auc_cutoff = 0.8)
}
\arguments{
\item{object}{A dataframe-like data object containing log-metabolite intensity values, with columns corresponding to metabolites and must containing the group column, and the rows corresponding to the samples}

\item{out_dir}{the out directory}

\item{seed}{the seed}

\item{auc_cutoff}{the cut off of auc that be filtered}
}
\value{
test
}
\description{
Feature selection using logistic regression
}
\examples{
library(MNet)

}
