% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_xgboost.R
\name{ML_xgboost}
\alias{ML_xgboost}
\title{feature selection in XGBoost}
\usage{
ML_xgboost(mydata)
}
\arguments{
\item{mydata}{the data}
}
\value{
test
}
\description{
feature selection in XGBoost
}
\examples{
library(dplyr)
mydata_t <- mydata \%>\%
  t() \%>\%
  as.data.frame()
# the group information must be tumor and normal
mydata_t$group <- group
result <- ML_xgboost(mydata_t)
}
