% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlimma.R
\name{mlimma}
\alias{mlimma}
\title{Differential Metabolite analysis by limma}
\usage{
mlimma(object, group)
}
\arguments{
\item{object}{object A dataframe-like data object containing raw metabolite intensity values, with rows corresponding to metabolites, and the columns corresponding to the samples}

\item{group}{the data's group}
}
\value{
A dataframe is same with limma's result
}
\description{
Differential Metabolite analysis by limma
}
\examples{
result <- mlimma(meta_dat,group)
}
