% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myscale.R
\name{myscale}
\alias{myscale}
\title{Scale the data}
\usage{
myscale(mydata, method = "log_zscore")
}
\arguments{
\item{mydata}{the data that row is the metabolites, and the column is the sample}

\item{method}{default is "log_zscore","raw_zscore" is alternative}
}
\value{
test
}
\description{
Scale the data
}
\examples{
mydata_scale <- myscale(mydata)
}
