% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayAnalysis.R
\name{PathwayAnalysis}
\alias{PathwayAnalysis}
\title{the KEGG pathway analysis which includes the extended pathway analysis of gene and metabolites}
\usage{
PathwayAnalysis(
  name,
  out = "Extended",
  p_cutoff = 0.05,
  noverlap_cutoff = 0,
  test = c("hypergeo", "fisher", "binomial")
)
}
\arguments{
\item{name}{The genes' or the metabolites' names which to analysis pathway}

\item{out}{The pathway type for gene or metabolite,or extended pathway included genes and metabolites,default is "extended",alternative is "metabolite" and "gene"}

\item{p_cutoff}{p_cutoff used to declare the significant terms. By default, it is set to 0.05}

\item{noverlap_cutoff}{noverlap_cutoff used to declare the number of overlap. By default, it is set to 0}

\item{test}{the test statistic used. It can be "fisher" for using fisher's exact test, "hypergeo" for using hypergeometric test, or "binomial" for using binomial test. Fisher's exact test is to test the independence between gene group (genes belonging to a group or not) and gene annotation (genes annotated by a term or not), and thus compare sampling to the left part of background (after sampling without replacement). Hypergeometric test is to sample at random (without replacement) from the background containing annotated and non-annotated genes, and thus compare sampling to background. Unlike hypergeometric test, binomial test is to sample at random (with replacement) from the background with the constant probability. In terms of the ease of finding the significance, they are in order: hypergeometric test > fisher's exact test > binomial test. In other words, in terms of the calculated p-value, hypergeometric test < fisher's exact test < binomial test}
}
\value{
test
}
\description{
the KEGG pathway analysis which includes the extended pathway analysis of gene and metabolites
}
\examples{
name <- c("C15973","C16254","MDH1")
result <- PathwayAnalysis(name,out="Extended",p_cutoff=0.05,noverlap_cutoff=0,test="hypergeo")
name <- "C15973"
result <- PathwayAnalysis(name,out="metabolite",p_cutoff=0.05,noverlap_cutoff=0,test="hypergeo")
name <- "MDH1"
result <- PathwayAnalysis(name,out="gene",p_cutoff=0.05,noverlap_cutoff=0,test="hypergeo")
}
