% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayinfo.R
\name{pathwayinfo}
\alias{pathwayinfo}
\title{get the gene and the enzyme in the pathway}
\usage{
pathwayinfo(pathwayid)
}
\arguments{
\item{pathwayid}{the pathway id, hsa or name}
}
\value{
test
}
\description{
get the gene and the enzyme in the pathway
}
\examples{
result <- pathwayinfo("hsa00630")
result <- pathwayinfo("Glyoxylate and dicarboxylate metabolism")
}
