% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayinfo.R
\name{pathwayinfo}
\alias{pathwayinfo}
\title{get the gene and the enzyme in the pathway}
\usage{
pathwayinfo(pathwayid)
}
\arguments{
\item{pathwayid}{the pathway kegg id}
}
\value{
test
}
\description{
get the gene and the enzyme in the pathway
}
\examples{
result <- pathwayinfo("hsa00630")
}
