% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdnet.R
\name{pdnet}
\alias{pdnet}
\title{The important network of the metabolite and the gene expression}
\usage{
pdnet(metabolite_data, gene_data, diff_info, nsize = 10)
}
\arguments{
\item{metabolite_data}{the row is the metabolite's kegg id, and the column is the sample}

\item{gene_data}{the row is the gene symbol, and the column is the sample}

\item{diff_info}{the differential genes and metabolites information}

\item{nsize}{the desired number of nodes constrained to the resulting subgraph}
}
\value{
a figure
}
\description{
The important network of the metabolite and the gene expression
}
\examples{
load("/Users/guituantuan/Desktop/R_packages/MNet/R/test.data")
pdf("test.pdf")
pdnet(metabolite_data,gene_data,diff_info,nsize=10)
dev.off()
}
