% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdnet.R
\name{pdnet}
\alias{pdnet}
\title{The subnetwork's extraction}
\usage{
pdnet(diff_metabolite, diff_gene, nsize = 10)
}
\arguments{
\item{diff_metabolite}{the differential metabolite information, the name 's kegg id}

\item{diff_gene}{the differential genes information}

\item{nsize}{the desired number of nodes constrained to the resulting subgraph}
}
\value{
a figure
}
\description{
The subnetwork's extraction
}
\examples{
library(dplyr)

names(diff_meta)[4]  <- "p_value"
diff_metabolite <- diff_meta \%>\%
  dplyr::filter(adj.P.Val < 0.01) \%>\%
  dplyr::filter(abs(logFC) > 1)
  
names(diff_gene)[4] <- "p_value"
diff_gene1 <- diff_gene \%>\%
  dplyr::filter(adj.P.Val < 0.01) \%>\%
  dplyr::filter(abs(logFC) > 1)
  
res <- pdnet(diff_meta, diff_gene, nsize = 100)
res

res <- pdnet(diff_metabolite[, 8], diff_gene1[1:500, 8])
res

}
