% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pPCA.R
\name{pPCA}
\alias{pPCA}
\title{Plot the PCA (Principal Component Analysis)}
\usage{
pPCA(object, group)
}
\arguments{
\item{object}{A dataframe-like data object containing raw metabolite intensity values, with rows corresponding to metabolites, and the columns corresponding to the samples}

\item{group}{the group}
}
\value{
Visualization of PCA
}
\description{
Plot the PCA (Principal Component Analysis)
}
\examples{
p_PCA <- pPCA(meta_dat, group)
p_PCA$p1
p_PCA$p2
p_PCA$p3

}
