% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pZscore.R
\name{pZscore}
\alias{pZscore}
\title{Plot the z-score figure}
\usage{
pZscore(mydata, group)
}
\arguments{
\item{mydata}{the row is the metabolites, and the column is the sample}

\item{group}{the sample's group information}
}
\value{
test
}
\description{
Plot the z-score figure
}
\examples{
plot_zscore <- pZscore(mydata,group)
}
