% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pZscore.R
\name{pZscore}
\alias{pZscore}
\title{Visualization of the z-score plot}
\usage{
pZscore(
  object,
  group,
  tumor_color = "#d53e4f",
  normal_color = "#7FC8A9",
  shape_size = 3,
  ysize = 5
)
}
\arguments{
\item{object}{A dataframe-like data object containing raw metabolite intensity values, with rows corresponding to metabolites, and the columns corresponding to the samples}

\item{group}{the sample's group information}

\item{tumor_color}{the color of the tumor group}

\item{normal_color}{the color of the normal group}

\item{shape_size}{the size of the point shape}

\item{ysize}{the size of the y-axis text}
}
\value{
Visualization of the z-score plot
}
\description{
Visualization of the z-score plot
}
\examples{
library(dplyr)

diff_result <- DM(2**meta_dat,group)
# filter the differential metabolites by default fold change >1.5 or < 1/1.5 ,fdr < 0.05 and VIP>1
diff_result_filter <- diff_result \%>\%
  dplyr::filter(Fold_change >1.3 | Fold_change < 1/1.3) \%>\%
  dplyr::filter(Padj_wilcox<0.1) \%>\%
  dplyr::filter(VIP>0.8)
meta_dat_diff <- meta_dat[rownames(meta_dat) \%in\% diff_result_filter$Name,]

p_zscore <- pZscore(meta_dat_diff,group)
p_zscore

}
