% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survCli.R
\name{survCli}
\alias{survCli}
\title{The survival analysis and visualization}
\usage{
survCli(clinical_survival)
}
\arguments{
\item{clinical_survival}{the data.frame must have the column name of status, time, group}
}
\value{
Figure
}
\description{
The survival analysis and visualization
}
\examples{
# names(aml)[3] ="group"
# plot = survCli(aml)
# plot

}
