% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survMet.R
\name{survMet}
\alias{survMet}
\title{The metabolites' survival plot
automatic classify the samples by each selected metabolites by mean or median, and then plot the survival}
\usage{
survMet(
  dat,
  metabolites,
  cluster_method = "mean",
  out_dir = "survival/metabolites/"
)
}
\arguments{
\item{dat}{the row is the sample, and the first 2 column is status and time, the rest of column
is the metabolites' value}

\item{metabolites}{the selected metabolites that used to classify the samples}

\item{cluster_method}{"mean" or "median",defalult is "mean"}

\item{out_dir}{the output directory,default is "survival/metabolites"}
}
\value{
test
}
\description{
The metabolites' survival plot
automatic classify the samples by each selected metabolites by mean or median, and then plot the survival
}
\examples{
metabolites <- c("C03819","C02918","C03916")
survMet(dat,metabolites,cluster_method="mean",out_dir="result/survival/metabolites/")
}
