% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgr.R
\name{xgr}
\alias{xgr}
\title{Plot the figure using XGR}
\usage{
xgr(metabolites_keggid, database)
}
\arguments{
\item{metabolites_keggid}{the metabolites's keggid}

\item{database}{the database used}
}
\value{
test
}
\description{
Plot the figure using XGR
}
\examples{
library(dplyr)
kegg_pathway_filter <- kegg_pathway \%>\%
dplyr::filter(!is.na(pathway_type)) \%>\%
  dplyr::select(c("ENTRY","PATHWAY"))

kegg_id_need <- c("C05984","C02494")
xgr_result <- xgr(kegg_id_need,kegg_pathway_filter)
xgr_result$output
xgr_result$gp
}
