% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgr.R
\name{xgr}
\alias{xgr}
\title{Plot the figure using XGR}
\usage{
xgr(
  metabolites_keggid,
  database,
  p_cutoff = 0.05,
  noverlap_cutoff = 0,
  test = c("hypergeo", "fisher", "binomial")
)
}
\arguments{
\item{metabolites_keggid}{the metabolites's keggid}

\item{database}{the database used}

\item{p_cutoff}{p_cutoff used to declare the significant terms. By default, it is set to 0.05}

\item{noverlap_cutoff}{noverlap_cutoff used to declare the number of overlap. By default, it is set to 0}

\item{test}{the test statistic used. It can be "fisher" for using fisher's exact test, "hypergeo" for using hypergeometric test, or "binomial" for using binomial test. Fisher's exact test is to test the independence between gene group (genes belonging to a group or not) and gene annotation (genes annotated by a term or not), and thus compare sampling to the left part of background (after sampling without replacement). Hypergeometric test is to sample at random (without replacement) from the background containing annotated and non-annotated genes, and thus compare sampling to background. Unlike hypergeometric test, binomial test is to sample at random (with replacement) from the background with the constant probability. In terms of the ease of finding the significance, they are in order: hypergeometric test > fisher's exact test > binomial test. In other words, in terms of the calculated p-value, hypergeometric test < fisher's exact test < binomial test}
}
\value{
test
}
\description{
Plot the figure using XGR
}
\examples{
library(dplyr)
kegg_pathway_filter <- kegg_pathway \%>\%
dplyr::filter(!is.na(pathway_type)) \%>\%
  dplyr::select(ENTRY,PATHWAY)

kegg_id_need <- c("C05984","C02494")
xgr_result <- xgr(kegg_id_need,kegg_pathway_filter,p_cutoff=1.1,noverlap_cutoff=0)
xgr_result$output
xgr_result$gp
}
